feat(nav): collapse sidebar to 5 entries
Moves four affordances out of the top-level sidebar:
- Inbox  -> bell icon top-right (rewired from /api/notifications to /api/inbox)
- Settings -> user-menu popover + Cmd/Ctrl+, shortcut
- Library -> /settings/library (with redirect for deep links)
- Dashboard -> click the Deft logo

Bell peek renders the same unified inbox feed (notifications + DM unread + pending approvals) used by /inbox; approval rows in the peek are passive - clicking routes to /inbox?tab=approvals rather than executing inline.

Sidebar nav goes from 9 entries to 5: Notes, Calendar, Chat, Tasks, Knowledge.
