--- apps/web/src/app/(app)/library/layout.tsx
@@ -1,5 +1,0 @@
-export const metadata = { title: 'Library' };
-
-export default function LibraryLayout({ children }: { children: React.ReactNode }) {
-  return children;
-}
