+++ apps/web/src/app/(app)/settings/library/layout.tsx
@@ -1,0 +1,4 @@
+export const metadata = { title: 'Library' };
+export default function Layout({ children }: { children: React.ReactNode }) {
+  return children;
+}
